+++ src/reducers/genres.js
@@ -1,0 +1,14 @@
+/**
+ * Used to store the genres.
+
+ */
+const genres = (state = [], action) => {
+  switch (action.type) {
+    case "SET_GENRES":
+      return action.payload;
+    default:
+      return state;
+  }
+};
+
+export default genres;
